+++ test/Big Integer.ts
@@ -1,6 +1,6 @@
 import { runTest } from './helpers'
 
-runTest<number, number>('BigInt', (test) => {
+runTest<number, number>('BigInteger', (test) => {
 	describe('validate', () => {
 		test.validate(1, true, 1)
 		test.validate('1', true, 1)
